refactor(database): extract pending state updates module
